tools/mc_texture_gen/requirements.txt
requests
pillow
rembg
